refactor: custom user build config
