+++ docs/en-US/development/plugin-ds.md
@@ -12,6 +12,7 @@ An empty plugin can be created quickly with `yarn pm create my-plugin`, with the
   |- package.json # plugin package information
   |- server.d.ts
   |- server.js
+  |- build.config.ts # or `build.config.js`, modify configuration
 ```
 
 The tutorial for `/src/server` refers to the [server](./server) section, and the tutorial for `/src/client` refers to the [client](./client) section.
@@ -19,7 +20,9 @@ The tutorial for `/src/server` refers to the [server](./server) section, and the
 If you want to customize the packaging configuration, you can create a `config.js` file in the root directory, with the following content:
 
 ```js
-module.exports = {
+import { defineConfig } from '@nocobase/build';
+
+export default defineConfig({
   modifyViteConfig: (config) => {
     // vite is used to package the `src/client` side code
 
@@ -32,5 +35,11 @@ module.exports = {
     // Modify the tsup configuration, for more information, please refer to: https://tsup.egoist.dev/#using-custom-configuration
     return config
   },
-}
+  beforeBuild: (log) => {
+    // The callback function before the build starts, you can do some operations before the build starts
+  },
+  afterBuild: (log: PkgLog) => {
+    // The callback function after the build is completed, you can do some operations after the build is completed
+  };
+})
 ```
