+++ docs/zh-CN/development/plugin-ds.md
@@ -12,7 +12,7 @@
   |- package.json   # 插件包信息
   |- server.d.ts
   |- server.js
-  |- config.js  # 自定义打包配置
+  |- build.config.ts # 或者 `build.config.js` ，用于修改打包配置，实现自定义逻辑
 ```
 
 `/src/server` 的教程参考 [服务端](./server) 章节，`/src/client` 的教程参考 [客户端](./client) 章节。
@@ -20,7 +20,9 @@
 如果你想要自定义打包配置，可以在根目录下创建 `config.js` 文件，内容如下：
 
 ```js
-module.exports = {
+import { defineConfig } from '@nocobase/build';
+
+export default defineConfig({
   modifyViteConfig: (config) => {
     // vite 是用来打包 `src/client` 端代码的
 
@@ -33,5 +35,11 @@ module.exports = {
     // 修改 tsup 配置，具体可参考：https://tsup.egoist.dev/#using-custom-configuration
     return config
   },
-}
+  beforeBuild: (log) => {
+    // 构建开始前的回调函数，可以在构建开始前做一些操作
+  },
+  afterBuild: (log: PkgLog) => {
+    // 构建完成后的回调函数，可以在构建完成后做一些操作
+  };
+});
 ```
